.ipynb_checkpoints/requirements-checkpoint.txt
tensorflow==2.16.1
gdown==4.6.0
pandas==1.5.3
gunicorn==20.1.0
matplotlib==3.8.4
streamlit== 1.34.0
Pillow==10.3.0
numpy==1.26.4
